[Does not work] First attempt at stereo recording
This commit is a (failed) first attempt at recording the uplink and
downlink audio streams as separate channels.

While the code works in theory, the proprietary `audio.primary.gs101.so`
audio driver on the Google Pixel 6 Pro does not work properly when
creating two AudioRecord instances for VOICE_UPLINK and VOICE_DOWNLINK.
Everything appears to work, but both instance produce audio data for the
downlink audio. This is not possible for BCR to work around.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>